Remove the barrel file from @markprompt/core in favor of exports
