+++ packages/core/src/constants.ts
@@ -1,0 +1,7 @@
+import { BaseOptions } from './types.js';
+
+export const DEFAULT_OPTIONS = {
+  apiUrl: 'https://api.markprompt.com',
+} satisfies BaseOptions;
+
+export type { BaseOptions };
