+++ packages/core/src/feedback.ts
@@ -1,9 +1,11 @@
 import defaults from 'defaults';
 
-import { DEFAULT_OPTIONS } from './index.js';
-import type { PromptFeedback, BaseOptions } from './types.js';
+import { DEFAULT_OPTIONS } from './constants.js';
+import type { BaseOptions } from './types.js';
 
-export type CSAT = 0 | 1 | 2 | 3 | 4 | 5;
+export interface PromptFeedback {
+  vote: '1' | '-1';
+}
 
 export interface SubmitFeedbackBody {
   /** Prompt feedback */
@@ -78,6 +80,8 @@ export async function submitFeedback(
     }
   }
 }
+
+export type CSAT = 0 | 1 | 2 | 3 | 4 | 5;
 
 export interface SubmitCSATBody {
   /** Thread id */
