+++ packages/core/src/types.ts
@@ -1,66 +1,3 @@
-import type { DocSearchHit } from './docsearch.js';
-
-export type {
-  Chat,
-  ChatCompletion,
-  ChatCompletionAssistantMessageParam,
-  ChatCompletionChunk,
-  ChatCompletionFunctionMessageParam,
-  ChatCompletionMessage,
-  ChatCompletionMessageParam,
-  ChatCompletionMessageToolCall,
-  ChatCompletionSystemMessageParam,
-  ChatCompletionTool,
-  ChatCompletionToolChoiceOption,
-  ChatCompletionToolMessageParam,
-  ChatCompletionUserMessageParam,
-} from 'openai/resources/chat/index.js';
-
-export type RequiredKeys<T, K extends keyof T> = Required<Pick<T, K>> &
-  Omit<T, K>;
-
-type ArrayToUnion<T> = T extends (infer U)[]
-  ? U
-  : T extends readonly (infer U)[]
-    ? U
-    : never;
-
-export const CHAT_COMPLETIONS_MODELS = [
-  'claude-3-5-sonnet-20240620',
-  'gpt-3.5-turbo',
-  'gpt-4-1106-preview',
-  'gpt-4-32k',
-  'gpt-4-turbo-preview',
-  'gpt-4',
-  'gpt-4o',
-  'gpt-4o-mini',
-  'meta-llama-3-70b-instruct',
-  'meta-llama-3-8b-instruct',
-  'mixtral-8x7b-instruct-v0.1',
-] as const;
-
-export type ChatCompletionsModel = ArrayToUnion<typeof CHAT_COMPLETIONS_MODELS>;
-
-export const COMPLETIONS_MODELS = [
-  'ada',
-  'babbage',
-  'curie',
-  'davinci',
-  'text-ada-001',
-  'text-babbage-001',
-  'text-curie-001',
-  'text-davinci-002',
-  'text-davinci-003',
-] as const;
-
-export type CompletionsModel = ArrayToUnion<typeof COMPLETIONS_MODELS>;
-
-export const EMBEDDINGS_MODEL = 'text-embedding-ada-002' as const;
-
-export type EmbeddingsModel = typeof EMBEDDINGS_MODEL;
-
-export type Model = ChatCompletionsModel | CompletionsModel | EmbeddingsModel;
-
 export type SourceType =
   | 'github'
   | 'motif'
@@ -80,22 +17,6 @@ export interface Source {
      */
     name?: string;
   };
-}
-
-export interface SearchResult extends SearchResultSection {
-  /**
-   * Reference file.
-   */
-  file: FileReferenceFileData;
-  /**
-   * Match type.
-   */
-  matchType: 'title' | 'leadHeading' | 'content';
-}
-
-export interface SearchResultSection extends FileSectionReferenceSectionData {
-  content?: string;
-  snippet?: string;
 }
 
 export interface FileSectionReference extends FileSectionReferenceSectionData {
@@ -136,37 +57,6 @@ export interface FileReferenceFileData {
    * File source.
    */
   source: Source;
-}
-
-export interface SearchResultsResponse {
-  debug?: unknown;
-  data: SearchResult[];
-}
-
-export type AlgoliaDocSearchHit = DocSearchHit;
-
-export interface AlgoliaDocSearchResultsResponse {
-  hits: AlgoliaDocSearchHit[];
-}
-
-export interface PromptFeedback {
-  vote: '1' | '-1';
-}
-
-export interface ChatCompletionMetadata {
-  threadId?: string;
-  messageId?: string;
-  references?: FileSectionReference[];
-  /**
-   * @deprecated Use `messageId` instead.
-   */
-  promptId?: string;
-}
-
-export interface NoStreamingData {
-  text: string;
-  references?: FileSectionReference[];
-  responseId: string;
 }
 
 export interface BaseOptions {
