+++ packages/core/tsconfig.build.json
@@ -6,5 +6,5 @@
     "rootDir": "src/"
   },
   "include": ["src/"],
-  "exclude": ["src/*.test.ts"]
+  "exclude": ["src/**/*.test.ts"]
 }
