+++ packages/react/src/primitives/headless.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import type { FileSectionReference } from '@markprompt/core';
+import type { FileSectionReference } from '@markprompt/core/types';
 import { AccessibleIcon } from '@radix-ui/react-accessible-icon';
 import * as Dialog from '@radix-ui/react-dialog';
 import {
